+++ src/components/CurriculumComponents/CurricTimetablingNewView/CurricTimetablingNewView.test.tsx
@@ -102,7 +102,7 @@ describe("CurricTimetablingNewView", () => {
     const linkElement = getByRole("link");
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=30&spring=30&summer=30&year=1",
+      "units?autumn=30&mode=normal&spring=30&summer=30&year=1",
     );
   });
 
@@ -172,28 +172,28 @@ describe("CurricTimetablingNewView", () => {
     // Initial state
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=30&spring=30&summer=30&year=1",
+      "units?autumn=30&mode=normal&spring=30&summer=30&year=1",
     );
 
     // Change autumn lessons to 25
     fireEvent.change(inputs[0]!, { target: { value: "25" } });
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=25&spring=30&summer=30&year=1",
+      "units?autumn=25&mode=normal&spring=30&summer=30&year=1",
     );
 
     // Change spring lessons to 15
     fireEvent.change(inputs[1]!, { target: { value: "15" } });
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=25&spring=15&summer=30&year=1",
+      "units?autumn=25&mode=normal&spring=15&summer=30&year=1",
     );
 
     // Change summer lessons to 20
     fireEvent.change(inputs[2]!, { target: { value: "20" } });
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=25&spring=15&summer=20&year=1",
+      "units?autumn=25&mode=normal&spring=15&summer=20&year=1",
     );
   });
 
@@ -211,7 +211,7 @@ describe("CurricTimetablingNewView", () => {
     const linkElement = getByRole("link");
     expect(linkElement).toHaveAttribute(
       "href",
-      "units?autumn=20&spring=25&summer=35&year=1",
+      "units?autumn=20&mode=normal&spring=25&summer=35&year=1",
     );
   });
 
@@ -225,7 +225,7 @@ describe("CurricTimetablingNewView", () => {
     expect(replaceStateSpy).toHaveBeenCalledWith(
       {},
       "",
-      "/timetabling/new?autumn=30&spring=30&summer=30&year=1",
+      "/timetabling/new?autumn=30&mode=normal&spring=30&summer=30&year=1",
     );
   });
 
@@ -239,7 +239,7 @@ describe("CurricTimetablingNewView", () => {
     expect(replaceStateSpy).toHaveBeenCalledWith(
       {},
       "",
-      "/timetabling/new?autumn=30&spring=30&summer=30&year=1",
+      "/timetabling/new?autumn=30&mode=normal&spring=30&summer=30&year=1",
     );
   });
 });